feat(pn-junction): 1D Sze permittivity, segmented optical waveguide
Consolidate stack/junction.py + stack/doping.py into stack/pn_junction.py
and the three pn-junction test files into tests/common/test_pn_junction.py.

Add the 1D free-carrier plasma-dispersion model (carrier_profile_1d,
epsilon_eff_relative, optical_params, junction_epsilon_profile) with
SI-correct units, plus make_segmented_junction_profile (8+8 uniform
p_1..n_8 strips sampling the Sze permittivity at strip centres).

The palace_2d_twmzm optical run now uses the segmented rib with per-strip
free-carrier materials (Re(eps)->Permittivity, Im(eps)->Conductivity);
build_optical_cross_section() accepts per-region device_materials and
extra_materials to support it.
