+++ docs/api/common.md
@@ -40,7 +40,49 @@
 
 ## PN Junction
 
-Depletion model after Sze & Ng, *Physics of Semiconductor Devices*, ch. 2.
+Depletion model after Sze & Ng, *Physics of Semiconductor Devices*, ch. 2,
+plus the 1D free-carrier plasma-dispersion model for the complex optical
+permittivity and the doping-profile geometry builders.
+
+::: gsim.common.stack.PNJunctionConfig
+    options:
+      show_source: false
+
+::: gsim.common.stack.make_pn_junction_profile
+    options:
+      show_source: false
+
+::: gsim.common.stack.make_segmented_junction_profile
+    options:
+      show_source: false
+
+::: gsim.common.stack.make_doping_profile
+    options:
+      show_source: false
+
+::: gsim.common.stack.junction_epsilon_profile
+    options:
+      show_source: false
+
+::: gsim.common.stack.carrier_profile_1d
+    options:
+      show_source: false
+
+::: gsim.common.stack.epsilon_eff_relative
+    options:
+      show_source: false
+
+::: gsim.common.stack.optical_params
+    options:
+      show_source: false
+
+::: gsim.common.stack.refractive_index
+    options:
+      show_source: false
+
+::: gsim.common.stack.drude_relaxation_times
+    options:
+      show_source: false
 
 ::: gsim.common.stack.PNJunctionConfig
     options:
