+++ src/gsim/common/stack/__init__.py
@@ -23,21 +23,12 @@ from pathlib import Path
 import gdsfactory as gf
 import yaml
 
-from gsim.common.stack.doping import make_doping_profile, make_pn_junction_profile
 from gsim.common.stack.extractor import (
     Layer,
     LayerStack,
     ValidationResult,
     extract_from_pdk,
     extract_layer_stack,
-)
-from gsim.common.stack.junction import (
-    PNJunctionConfig,
-    built_in_voltage,
-    depletion_extents,
-    depletion_width,
-    junction_capacitance_per_area,
-    select_junction_mode,
 )
 from gsim.common.stack.materials import (
     MATERIALS_DB,
@@ -56,6 +47,30 @@ from gsim.common.stack.materials import (
 from gsim.common.stack.overlays import (
     load_overlay,
     merge_overlay,
+)
+from gsim.common.stack.pn_junction import (
+    M_CE_STAR,
+    M_CH_STAR,
+    MU_N_CM2_VS,
+    MU_P_CM2_VS,
+    NI_SI_300K_CM3,
+    SIGMA_NEGLIGIBLE_SM,
+    PNJunctionConfig,
+    built_in_voltage,
+    carrier_profile_1d,
+    default_eps_bg_rel,
+    depletion_extents,
+    depletion_width,
+    drude_relaxation_times,
+    epsilon_eff_relative,
+    junction_capacitance_per_area,
+    junction_epsilon_profile,
+    make_doping_profile,
+    make_pn_junction_profile,
+    make_segmented_junction_profile,
+    optical_params,
+    refractive_index,
+    select_junction_mode,
 )
 from gsim.common.stack.visualization import (
     StackLayer,
@@ -173,6 +188,12 @@ def load_stack_yaml(yaml_path: str | Path) -> LayerStack:
 
 __all__ = [
     "MATERIALS_DB",
+    "MU_N_CM2_VS",
+    "MU_P_CM2_VS",
+    "M_CE_STAR",
+    "M_CH_STAR",
+    "NI_SI_300K_CM3",
+    "SIGMA_NEGLIGIBLE_SM",
     "DispersionModel",
     "Layer",
     "LayerStack",
@@ -185,24 +206,32 @@ __all__ = [
     "ValidationResult",
     "ValidityRange",
     "built_in_voltage",
+    "carrier_profile_1d",
+    "default_eps_bg_rel",
     "depletion_extents",
     "depletion_width",
+    "drude_relaxation_times",
+    "epsilon_eff_relative",
     "extract_from_pdk",
     "extract_layer_stack",
     "get_material_properties",
     "get_stack",
     "junction_capacitance_per_area",
+    "junction_epsilon_profile",
     "load_overlay",
     "load_stack_yaml",
     "make_doped_material",
     "make_doped_materials",
     "make_doping_profile",
     "make_pn_junction_profile",
+    "make_segmented_junction_profile",
     "merge_overlay",
+    "optical_params",
     "parse_layer_stack",
     "plot_stack",
     "print_stack",
     "print_stack_table",
+    "refractive_index",
     "resolve_material_at_wavelength",
     "select_junction_mode",
     "should_enable_dispersion",
